+++ apps/frontend/src/app/3_organisms/TransactionStepDialog/TransactionStepDialog.types.ts
@@ -50,6 +50,9 @@ export enum TransactionType {
   signTypedData = 'signTypedData',
   signTransaction = 'signTransaction',
   signTransactionData = 'signTransactionData',
+  /**
+   * @deprecated Use `signTypedData` instead.
+   */
   signPermit = 'signPermit',
 }
 
@@ -64,9 +67,12 @@ export type SignTypedDataRequest = {
   signer: JsonRpcSigner;
   domain: TypedDataDomain;
   types: Record<string, Array<TypedDataField>>;
-  value: Record<string, any>;
+  values: Record<string, any>;
 };
 
+/**
+ * @deprecated Use TypedDataTransactionRequest instead
+ */
 export type SignPermitRequest = {
   type: TransactionType.signPermit;
   signer: JsonRpcSigner;
@@ -115,5 +121,5 @@ export enum TransactionReceiptStatus {
 export type TransactionReceipt = {
   status: TransactionReceiptStatus;
   request: TransactionRequest;
-  response?: string | PermitTransactionResponse;
+  response?: string;
 };