+++ apps/frontend/src/app/5_pages/ClaimLpPage/hooks/useClaimLp.ts
@@ -78,7 +78,7 @@ export const useClaimLp = (onComplete?: () => void) => {
         .catch(() => [])) as number[];
     }
 
-    return claims.filter(c => unclaimedIndexes.includes(c.index));
+    return claims.filter(c => unclaimedIndexes?.includes(c.index));
   }, [contract, findProofs]);
 
   const claim = useCallback(async () => {