+++ apps/frontend/src/app/5_pages/MarketMakingPage/MarketMakingPage.tsx
@@ -2,6 +2,7 @@ import React, { FC } from 'react';
 
 import { RSK_CHAIN_ID } from '../../../config/chains';
 
+import { CrocContextProvider } from '../../../contexts/CrocContext';
 import { useCurrentChain } from '../../../hooks/useChainStore';
 import { AmbientMarketMaking } from './components/AmbientMarketMaking/AmbientMarketMaking';
 import { MarketMaking } from './components/MarketMaking/MarketMaking';
@@ -10,7 +11,11 @@ const MarketMakingPage: FC = () => {
   const currentChainId = useCurrentChain();
 
   if (currentChainId !== RSK_CHAIN_ID) {
-    return <AmbientMarketMaking />;
+    return (
+      <CrocContextProvider>
+        <AmbientMarketMaking />
+      </CrocContextProvider>
+    );
   }
   return <MarketMaking />;
 };