+++ packages/contracts/src/contracts/assets/rsk.ts
@@ -118,4 +118,11 @@ export const rsk: Array<AssetDetails> = [
     decimals: 18,
     getIcon: async () => (await import('./icons/rsk/mynt')).default,
   },
+  {
+    symbol: 'POWA',
+    address: '0xfe8929d36ac789c562008abd59f5dd7e1eb1f5b6',
+    name: 'POWA•RANGERS•GO',
+    decimals: 18,
+    getIcon: async () => (await import('./icons/rsk/powa')).default,
+  },
 ];