+++ packages/contracts/src/contracts/protocol/bobTestnet.ts
@@ -6,9 +6,14 @@ export const bobTestnet: Record<string, AsyncContractConfigData> = {
     getAbi: async () => (await import('../../abis/staking.json')).default,
   },
   merkleDistributor: {
-    address: '0x0', // todo: update address
+    address: '0x0',
     getAbi: async () =>
       (await import('../../abis/merkleDistributor.json')).default,
+  },
+  stakingRewardsOs: {
+    address: '0x0',
+    getAbi: async () =>
+      (await import('../../abis/StakingRewardsOs.json')).default,
   },
   runeBridge: {
     address: '0x9d451acb7A7f98dc4F2d2A2aA1A0b0436f0Effdb',